Function tink::cryptofmt::output_prefix

pub fn output_prefix(key: &Key) -> Result<Vec<u8>, TinkError>

Generate the prefix of ciphertexts produced by the crypto primitive obtained from key. The prefix can be either empty (for RAW-type prefix), or consists of a 1-byte indicator of the type of the prefix, followed by 4 bytes of the key ID in big endian encoding.