Module tink::keyset::insecure

Module for test code methods to read or write cleartext keyset material.

Functions

keyset_material
Return the key material contained in a Handle.

new_handle
Create a new instance of Handle using the given Keyset.

read
Create a Handle from a cleartext keyset obtained via r.

write
Exports the keyset from h to the given writer w without encrypting it. Storing secret key material in an unencrypted fashion is dangerous. If feasible, you should use super::Handle::write() instead.