Module tink::subtle

Common methods needed in subtle implementations.

Modules

random
Utilities for random data.

Enums

HashFunc
Hash function object.

Functions

compute_hash
Calculate a hash of the given data using the given hash function.

compute_hkdf
Extract a pseudorandom key.

constant_time_compare
Compare two slices in constant time. Return true if they are equal, false otherwise.

get_hash_digest_size
Return the digest size of the specified hash algorithm.

get_hash_func
Return the corresponding hash function of the given hash.